Test files added to project
These Arduino Sketches are test projects that I put together to test my
wiring harness as well as make sure that the conversion equations are
working properly. MultiAnalogTestHeader uses a header file and a c++
class that is created in the arduino sketch. The
MulitAnalogeRangeFinder.h and MulitRangeFinder.cpp can be used to do the
range calculations.